app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
cryptography
python-dotenv
